+++ app/.storybook/preview.tsx
@@ -1,41 +1,14 @@
-import "@foxglove-studio/app/styles/global.scss";
-import "./styles.scss";
 import waitForFonts from "@foxglove-studio/app/util/waitForFonts";
 import { withScreenshot } from "storycap";
 import { withMockSubscribeToNewsletter } from "./__mocks__/subscribeToNewsletter";
 import { Story, StoryContext } from "@storybook/react";
 import ThemeProvider from "@foxglove-studio/app/theme/ThemeProvider";
-import signal from "@foxglove-studio/app/shared/signal";
-import ScreenshotContext from "@foxglove-studio/app/stories/ScreenshotContext";
-import { useCallback, useRef, useState } from "react";
+
+import "@foxglove-studio/app/styles/global.scss";
+import "./styles.scss";
+import withScreenshotSignal from "@foxglove-studio/app/.storybook/withScreenshotSignal";
 
 let loaded = false;
-
-function withScreenshotSignal(Story: Story, { parameters }: StoryContext) {
-  const signalRef = useRef(signal());
-  const callCount = useRef(0);
-  const [error, setError] = useState<Error | undefined>(undefined);
-  parameters.screenshot.waitFor = signalRef.current;
-
-  const sceneReady = useCallback(() => {
-    if (callCount.current > 0) {
-      setError(new Error("withScreenshotSignal: called scene ready more than once"));
-      return;
-    }
-    ++callCount.current;
-    signalRef.current.resolve();
-  }, []);
-
-  if (error) {
-    throw error;
-  }
-
-  return (
-    <ScreenshotContext.Provider value={sceneReady}>
-      <Story />
-    </ScreenshotContext.Provider>
-  );
-}
 
 function withTheme(Story: Story, { parameters }: StoryContext) {
   return (
